Split dynamic and snapshot tree classes
