+++ ydb/core/kqp/runtime/scheduler/new/kqp_compute_actor.h
@@ -27,7 +27,7 @@ namespace NKikimr::NKqp::NScheduler {
         // Magic state function name to overload
         STATEFN(BaseStateFuncBody) {
             // TODO: account mailbox usage?
-            // we assume that exception handling is done in parents/descendants
+            // we assume that exceptions are handled in parents/descendants
             switch (ev->GetTypeRewrite()) {
                 hFunc(NActors::TEvents::TEvWakeup, TSchedulableComputeActorBase<TDerived>::Handle);
                 default:
